start work on the specs and define TendermintRpcClientExt
