+++ common/client-libs/validator-client/src/nyxd/cosmwasm_client/module_traits/feegrant/query.rs
@@ -138,7 +138,7 @@ pub trait FeegrantQueryClient: CosmWasmClient {
         };
 
         let res = self
-            .make_abci_query::<ProtoQueryAllowancesRequest, ProtoQueryAllowancesResponse>(
+            .make_abci_query_without_proof::<ProtoQueryAllowancesRequest, ProtoQueryAllowancesResponse>(
                 path,
                 req.into(),
             )
