+++ common/credential-proxy/src/shared_state/nyxd_client.rs
@@ -6,7 +6,7 @@ use crate::helpers::LockTimer;
 use nym_ecash_contract_common::msg::ExecuteMsg;
 use nym_validator_client::nyxd::contract_traits::NymContractsProvider;
 use nym_validator_client::nyxd::cosmwasm_client::types::ExecuteResult;
-use nym_validator_client::nyxd::{AccountId, Coin, Config, CosmWasmClient, NyxdClient};
+use nym_validator_client::nyxd::{AccountId, Coin, Config, NyxdClient, TendermintRpcClientExt};
 use nym_validator_client::{DirectSigningHttpRpcNyxdClient, nyxd};
 use std::ops::Deref;
 use std::sync::Arc;
