+++ nym-api/src/node_status_api/cache/refresher.rs
@@ -27,7 +27,8 @@ use nym_mixnet_contract_common::{NodeId, NymNodeDetails};
 use nym_task::ShutdownToken;
 use nym_topology::CachedEpochRewardedSet;
 use nym_validator_client::nyxd::module_traits::feegrant::query::FeegrantQueryClient;
-use nym_validator_client::nyxd::{AccountId, CosmWasmClient};
+use nym_validator_client::nyxd::AccountId;
+use nym_validator_client::rpc::TendermintRpcClientExt;
 use nym_validator_client::QueryHttpRpcNyxdClient;
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
