+++ openspec/changes/directory-retrieval-client/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-02
